requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
tensorflow==2.14.0
scikit-learn==1.3.2
pymongo==4.6.0
python-dotenv==1.0.0
pytz==2023.3